+++ crates/rust-client/src/store/web_store/js/schema.js
@@ -29,6 +29,7 @@ const Table = {
   BlockHeaders: "blockHeaders",
   PartialBlockchainNodes: "partialBlockchainNodes",
   Tags: "tags",
+  ForeignAccountCode: "foreignAccountCode",
 };
 
 const db = new Dexie(DATABASE_NAME);