+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Features
 
 * Added support for FPI in Web Client (#958).
+* Exposed `bech32` account IDs in Web Client (#978).
 
 ### Fixes